requirements.txt
PyQt5
PyPDF2
python-docx
docx2pdf
Pillow